Updated the SGEEE pending job algorithm.
WARNING - WARNING - WARNING - WARNING - WARNING - WARNING

This update changes the binary job structure spooled to disk.  This means
that if you are already running with the open source code, you will need
to delete all pending and active jobs before running with the new SGE
binaries. This is true if you are running either SGE or SGEEE.