EB-2001-12-10-1: Enhancem.: Memory optimization for zombie jobs
                            THE SPOOL FILE FORMAT FOR JOBS (CULL JOB STRUCTURE)
                            HAS BEEN CHANGED

WARNING - WARNING - WARNING - WARNING

This update changes the binary job structure spooled to                         disk.  This means that if you are already running with                          the open source code, you will need to delete all pending
and active jobs before running with the new SGE binaries.
This is true if you are running either SGE or SGEEE